Refactored to move login method to own module included in spec_helper. Updated spec files to use it.
