+++ spec/features/users/log_in_spec.rb
@@ -18,7 +18,7 @@ feature "Log in" do
   end
 
   it "can log in with valid name and password" do
-    # Use pre-existing user and custom method defined below
+    # Use pre-existing user and custom method defined in login_helper.rb
     login("lupe_howell@jacobsonnolan.info", "password1")
     expect(page).to have_content "Sign out"
     expect(page).to have_css(".logout")
@@ -26,7 +26,7 @@ feature "Log in" do
   end
 
   context "can't log in with missing or invalid credentials" do
-    it "can't log in without enetering an email address" do
+    it "can't log in without entering an email address" do
       click_button "Sign in"
       expect(page).to have_content "An email address required."
     end
@@ -46,15 +46,4 @@ feature "Log in" do
       expect(page).to have_content "Authentication failed."
     end
   end
-end
-
-# helper method to log in
-def login(email, password)
-  # Use css to find correct email address field input
-  # because there are two email address fields on the page
-  # and 'within' block doesn't work here
-  email_field = find(:css, "input[id$='email']")
-  email_field.fill_in with: email
-  fill_in "passwd", with: password
-  click_button "Sign in"
 end
